flask-ai-chatbot/requirements.txt
openai>=1.0.0
flask>=2.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-flask>=1.2.0